refactor(suite): handleProtocolRequest thunk
