+++ packages/suite/src/actions/suite/protocolActions.ts
@@ -1,4 +1,7 @@
 import { PROTOCOL } from './constants';
+import { getProtocolInfo, isProtocolScheme } from '@suite-utils/protocol';
+import { addToast } from '@suite-actions/notificationActions';
+import type { Dispatch } from '@suite-types';
 import type { PROTOCOL_SCHEME } from '@suite-constants/protocol';
 import type { SendFormState } from '@suite-reducers/protocolReducer';
 
@@ -18,7 +21,7 @@ export const fillSendForm = (shouldFill: boolean): ProtocolAction => ({
     payload: shouldFill,
 });
 
-export const saveCoinProtocol = (
+const saveCoinProtocol = (
     scheme: PROTOCOL_SCHEME,
     address: string,
     amount?: number,
@@ -26,6 +29,25 @@ export const saveCoinProtocol = (
     type: PROTOCOL.SAVE_COIN_PROTOCOL,
     payload: { scheme, address, amount },
 });
+
+export const handleProtocolRequest = (uri: string) => (dispatch: Dispatch) => {
+    const protocol = getProtocolInfo(uri);
+
+    if (protocol && isProtocolScheme(protocol.scheme)) {
+        const { scheme, amount, address } = protocol;
+
+        dispatch(saveCoinProtocol(scheme, address, amount));
+        dispatch(
+            addToast({
+                type: 'coin-scheme-protocol',
+                address,
+                scheme,
+                amount,
+                autoClose: false,
+            }),
+        );
+    }
+};
 
 export const resetProtocol = (): ProtocolAction => ({
     type: PROTOCOL.RESET,
